feat: move tva verification on client side
