+++ clients/routes.ts
@@ -132,7 +132,6 @@ const routes = {
       'https://staging.recherche-entreprises.api.gouv.fr/search',
     uniteLegale: 'https://api-annuaire-entreprises.data.gouv.fr/siren?q=',
   },
-  tva: 'https://rncs-proxy.api.gouv.fr/tva/',
 };
 
 export default routes;
